Android: avoid duplicated entries when reporting satellite info
It turns out that some devices can report the same satellite
(identified by a combination of satellite system and satellite id)
multuiple times in the same GNSS Status callback. The duplicated
devices seem to have just a different value of RSSI, otherwise
they are identical.

This behavior is visible at least on Oneplus devices.

This does not seem to be correct, and that can be a source of
bugs on the user side, if they expect the items in the list to
be unique.

Fix it by keeping track of all devices and simply discarding the
duplicates. We have no way to figure out which information is
newer, so simply keep the first one. In most cases it does not
matter anyway, as the satellite updates are received on a regular
basis.

Change-Id: I5c266d85cd65a8729174807dba747f8fee0ed2d0
Reviewed-by: Juha Vuolle <juha.vuolle@qt.io>
(cherry picked from commit 58bbf8fa5fdeb1c19d71cdab8830b9cf55672a5e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit dc6e0b8c31713ad80b5eee24218ac025fcd113eb)