QtPositioning: add docs for member-swap()
This patch adds the missing documentation for member-swap() methods of
various Qt Positioning classes. The patch uses the new \memberswap
macro.

All the member-swap() methods were added in Qt 6.2, when the module
was first reintroduced for Qt 6. So, pick this patch all the way down
to that version.

Task-number: QTBUG-129789
Pick-to: 6.8 6.5 6.2
Change-Id: Icfdaa6aa1984e816c425e776487e9e87a7a57e09
Reviewed-by: Marc Mutz <marc.mutz@qt.io>
Reviewed-by: Edward Welbourne <edward.welbourne@qt.io>